Added unicity checking, existance checking, creation/edition/deletion messages.
1) Checking problem type/device type/urgency status unicity;
2) Checking orders with problem type/urgency status before deleting;
3) Checking devices with device type before deleting;
4) Messages about creating/editing/deleting of problem type/device
type/urgency status/order.

+++ src/main/java/lannisters/devcor/controller/AdminController.java
@@ -25,6 +25,7 @@ import lannisters.devcor.entity.UrgencyStatus;
 import lannisters.devcor.mail.MailService;
 import lannisters.devcor.service.DeviceTypesService;
 import lannisters.devcor.service.DevicesService;
+import lannisters.devcor.service.OrdersService;
 import lannisters.devcor.service.PlayersService;
 import lannisters.devcor.service.ProblemTypesService;
 import lannisters.devcor.service.ReportService;
@@ -51,6 +52,9 @@ public class AdminController {
 
 	@Autowired
 	private RoomsService roomsService;
+
+	@Autowired
+	private OrdersService ordersService;
 
 	@Autowired
 	private ReportService reportService;
@@ -382,9 +386,20 @@ public class AdminController {
 	}
 
 	@RequestMapping(value = "/configuration/problemType/edit/id/{id}", method = RequestMethod.POST)
-	public String saveEditedProblemType(@ModelAttribute("problem") ProblemType problemType) {
-		problemTypesService.updateProblemType(problemType);
-		return "redirect:/configuration";
+	public String saveEditedProblemType(@ModelAttribute("problem") ProblemType problemType, @PathVariable("id") int id,
+			RedirectAttributes redirectAttributes) {
+		String page = null;
+		if (problemTypesService.checkProblemTypeExistence(problemType)
+				&& problemTypesService.getProblemTypeByTitle(problemType.getProblemType())
+						.getProblemTypeId() != problemType.getProblemTypeId()) {
+			redirectAttributes.addFlashAttribute("unique", "Problem type with this title exist!");
+			page = "redirect:/configuration/problemType/edit/id/" + id;
+		} else {
+			problemTypesService.updateProblemType(problemType);
+			redirectAttributes.addFlashAttribute("message", "Problem type edited successfully!");
+			page = "redirect:/configuration";
+		}
+		return page;
 	}
 
 	@RequestMapping(value = "/configuration/problemType/add", method = RequestMethod.GET)
@@ -395,14 +410,28 @@ public class AdminController {
 	}
 
 	@RequestMapping(value = "/configuration/problemType/add", method = RequestMethod.POST)
-	public String saveProblemType(@ModelAttribute("problem") ProblemType problemType) {
-		problemTypesService.addProblemType(problemType);
-		return "redirect:/configuration";
+	public String saveProblemType(@ModelAttribute("problem") ProblemType problemType,
+			RedirectAttributes redirectAttributes) {
+		String page = null;
+		if (problemTypesService.checkProblemTypeExistence(problemType)) {
+			redirectAttributes.addFlashAttribute("unique", "Problem type with this title exist!");
+			page = "redirect:/configuration/problemType/add";
+		} else {
+			problemTypesService.addProblemType(problemType);
+			redirectAttributes.addFlashAttribute("message", "Problem type added successfully!");
+			page = "redirect:/configuration";
+		}
+		return page;
 	}
 
 	@RequestMapping(value = "/configuration/problemType/delete/id/{id}", method = RequestMethod.GET)
-	public String deleteProblemType(@PathVariable("id") int id) {
-		problemTypesService.deleteProblemType(id);
+	public String deleteProblemType(@PathVariable("id") int id, RedirectAttributes redirectAttributes) {
+		if (ordersService.checkProblemTypeOrders(id))
+			redirectAttributes.addFlashAttribute("error", "Unable to delete. Orders with this problem type exist!");
+		else {
+			problemTypesService.deleteProblemType(id);
+			redirectAttributes.addFlashAttribute("message", "Problem type deleted successfully!");
+		}
 		return "redirect:/configuration";
 	}
 
@@ -414,9 +443,19 @@ public class AdminController {
 	}
 
 	@RequestMapping(value = "/configuration/deviceType/edit/id/{id}", method = RequestMethod.POST)
-	public String saveEditedDeviceType(@ModelAttribute("deviceType") DeviceType deviceType) {
-		deviceTypesService.updateDeviceType(deviceType);
-		return "redirect:/configuration";
+	public String saveEditedDeviceType(@ModelAttribute("deviceType") DeviceType deviceType, @PathVariable("id") int id,
+			RedirectAttributes redirectAttributes) {
+		String page = null;
+		if (deviceTypesService.checkDeviceTypeExistence(deviceType) && deviceTypesService
+				.getDeviceTypeByTitle(deviceType.getDeviceType()).getDeviceTypeId() != deviceType.getDeviceTypeId()) {
+			redirectAttributes.addFlashAttribute("unique", "Device type with this title exist!");
+			page = "redirect:/configuration/deviceType/edit/id/" + id;
+		} else {
+			deviceTypesService.updateDeviceType(deviceType);
+			redirectAttributes.addFlashAttribute("message", "Device type edited successfully!");
+			page = "redirect:/configuration";
+		}
+		return page;
 	}
 
 	@RequestMapping(value = "/configuration/deviceType/add", method = RequestMethod.GET)
@@ -427,14 +466,28 @@ public class AdminController {
 	}
 
 	@RequestMapping(value = "/configuration/deviceType/add", method = RequestMethod.POST)
-	public String saveDeviceType(@ModelAttribute("deviceType") DeviceType deviceType) {
-		deviceTypesService.addDeviceType(deviceType);
-		return "redirect:/configuration";
+	public String saveDeviceType(@ModelAttribute("deviceType") DeviceType deviceType,
+			RedirectAttributes redirectAttributes) {
+		String page = null;
+		if (deviceTypesService.checkDeviceTypeExistence(deviceType)) {
+			redirectAttributes.addFlashAttribute("unique", "Device type with this title exist!");
+			page = "redirect:/configuration/deviceType/add";
+		} else {
+			deviceTypesService.addDeviceType(deviceType);
+			redirectAttributes.addFlashAttribute("message", "Device type added successfully!");
+			page = "redirect:/configuration";
+		}
+		return page;
 	}
 
 	@RequestMapping(value = "/configuration/deviceType/delete/id/{id}", method = RequestMethod.GET)
-	public String deleteDeviceType(@PathVariable("id") int id) {
-		deviceTypesService.deleteDeviceType(id);
+	public String deleteDeviceType(@PathVariable("id") int id, RedirectAttributes redirectAttributes) {
+		if (devicesService.checkDeviceTypeDevices(id))
+			redirectAttributes.addFlashAttribute("error", "Unable to delete. Devices with this device type exist!");
+		else {
+			deviceTypesService.deleteDeviceType(id);
+			redirectAttributes.addFlashAttribute("message", "Device type deleted successfully!");
+		}
 		return "redirect:/configuration";
 	}
 
@@ -446,9 +499,20 @@ public class AdminController {
 	}
 
 	@RequestMapping(value = "/configuration/urgStatus/edit/id/{id}", method = RequestMethod.POST)
-	public String saveEditedUrgStatus(@ModelAttribute("urgStatus") UrgencyStatus urgencyStatus) {
-		urgencyStatusesService.updateUrgencyStatus(urgencyStatus);
-		return "redirect:/configuration";
+	public String saveEditedUrgStatus(@ModelAttribute("urgStatus") UrgencyStatus urgencyStatus,
+			@PathVariable("id") int id, RedirectAttributes redirectAttributes) {
+		String page = null;
+		if (urgencyStatusesService.checkUrgencyStatusExistence(urgencyStatus)
+				&& urgencyStatusesService.getUrgencyStatusByTitle(urgencyStatus.getUrgencyStatus())
+						.getUrgencyStatusId() != urgencyStatus.getUrgencyStatusId()) {
+			redirectAttributes.addFlashAttribute("unique", "Urgency status with this title exist!");
+			page = "redirect:/configuration/urgStatus/edit/id/" + id;
+		} else {
+			urgencyStatusesService.updateUrgencyStatus(urgencyStatus);
+			redirectAttributes.addFlashAttribute("message", "Urgency status edited successfully!");
+			page = "redirect:/configuration";
+		}
+		return page;
 	}
 
 	@RequestMapping(value = "/configuration/urgStatus/add", method = RequestMethod.GET)
@@ -459,14 +523,28 @@ public class AdminController {
 	}
 
 	@RequestMapping(value = "/configuration/urgStatus/add", method = RequestMethod.POST)
-	public String saveUrgStatus(@ModelAttribute("urgStatus") UrgencyStatus urgencyStatus) {
-		urgencyStatusesService.addUrgencyStatus(urgencyStatus);
-		return "redirect:/configuration";
+	public String saveUrgStatus(@ModelAttribute("urgStatus") UrgencyStatus urgencyStatus,
+			RedirectAttributes redirectAttributes) {
+		String page = null;
+		if (urgencyStatusesService.checkUrgencyStatusExistence(urgencyStatus)) {
+			redirectAttributes.addFlashAttribute("unique", "Urgency status with this title exist!");
+			page = "redirect:/configuration/urgStatus/add";
+		} else {
+			urgencyStatusesService.addUrgencyStatus(urgencyStatus);
+			redirectAttributes.addFlashAttribute("message", "Urgency status added successfully!");
+			page = "redirect:/configuration";
+		}
+		return page;
 	}
 
 	@RequestMapping(value = "/configuration/urgStatus/delete/id/{id}", method = RequestMethod.GET)
-	public String deleteUrgStatus(@PathVariable("id") int id) {
-		urgencyStatusesService.deleteUrgencyStatus(id);
+	public String deleteUrgStatus(@PathVariable("id") int id, RedirectAttributes redirectAttributes) {
+		if (ordersService.checkUrgencyStatusOrders(id))
+			redirectAttributes.addFlashAttribute("error", "Unable to delete. Orders with this urgency status exist!");
+		else {
+			urgencyStatusesService.deleteUrgencyStatus(id);
+			redirectAttributes.addFlashAttribute("message", "Urgency status deleted successfully!");
+		}
 		return "redirect:/configuration";
 	}
 }
